requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
Pillow==7.2.0
psycopg2-binary==2.9.4
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
